requirements.txt
flask>=3.0.0
yfinance>=0.2.40
anthropic>=0.40.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
pandas>=2.0.0
gunicorn>=21.2.0
